Refactor - Modularized the get_caption function
In order to be able to use part of it again in the futur
Sankey's diagram captions.
