+++ survey/exporter/tex/question2tex_chart.py
@@ -8,11 +8,9 @@ import logging
 
 from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
-from django.utils.translation import ungettext
 from future import standard_library
 
 from survey.exporter.tex.question2tex import Question2Tex
-from survey.models.question import Question
 
 standard_library.install_aliases()
 
@@ -125,64 +123,9 @@ class Question2TexChart(Question2Tex):
         else:
             return ""
 
-    def get_caption(self):
-        """ Return a caption with an appropriate description of the chart. """
-        caption = "{} ".format(_("Repartition of answers"))
-        if self.min_cardinality > 0:
-            caption += "{} {} ".format(
-                _("with"),
-                ungettext(
-                    "%(min_cardinality)d respondants or more",
-                    "%(min_cardinality)d respondant or more",
-                    self.min_cardinality
-                ) % {'min_cardinality': self.min_cardinality, }
-            )
-        if self.filter:
-            caption += "{} ".format(_("excluding"))
-            for i, excluded in enumerate(self.filter):
-                excluded = Question2Tex.get_clean_answer(excluded)
-                caption += "'{}', ".format(excluded)
-                if len(self.filter) >= 2 and i == len(self.filter) - 2:
-                    caption += "{} ".format(_("and"))
-            caption = "{} ".format(caption[:-2])
-        caption += "%s '%s' " % (
-            _("for the question"), Question2Tex.html2latex(self.question.text)
-        )
-        if self.group_together:
-            if self.cardinality is None:
-                loop_dict = self.group_together
-            else:
-                # Looping only on the value really used in the answers
-                loop_dict = self.cardinality
-            has_and = False
-            for key in loop_dict:
-                values = self.group_together.get(key)
-                if values is None:
-                    # group_together does not contain every answers
-                    continue
-                standardized_values = Question.standardize_list(
-                    values, self.group_by_letter_case, self.group_by_slugify
-                )
-                standardized_key = Question.standardize(
-                    key, self.group_by_letter_case, self.group_by_slugify
-                )
-                relevant_values = [v for v in standardized_values
-                                   if v != standardized_key]
-                if not relevant_values:
-                    # If there is no relevant value the group_together was just
-                    # a placeholder ex Yes for [yes Yës yEs]
-                    continue
-                # We duplicate the translations so makemessage find it
-                caption += "with '{}' standing for ".format(key)
-                for value in values:
-                    caption += "'{}' {} ".format(value, _("or"))
-                caption = caption[:-len("{} ".format(_("or")))]
-                has_and = True
-                caption += "{} ".format(_("and"))
-            if has_and:
-                # We remove the final "and " if there is one
-                caption = caption[:-len("{} ".format(_("and")))]
-        return "{}.".format(caption[:-1])
+    def get_caption_specifics(self):
+        return "%s '%s' " % (_("for the question"),
+                             Question2Tex.html2latex(self.question.text))
 
     def tex(self):
         """ Return a pfg-pie pie chart of a question.
